+++ ui/pages/confirmations/components/confirm/header/wallet-initiated-header.test.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { DefaultRootState } from 'react-redux';
+import { getMockTokenTransferConfirmState } from '../../../../../../test/data/confirmations/helper';
+import { renderWithConfirmContextProvider } from '../../../../../../test/lib/confirmations/render-helpers';
+import configureStore from '../../../../../store/store';
+import { WalletInitiatedHeader } from './wallet-initiated-header';
+
+const render = (
+  state: DefaultRootState = getMockTokenTransferConfirmState({}),
+) => {
+  const store = configureStore(state);
+  return renderWithConfirmContextProvider(<WalletInitiatedHeader />, store);
+};
+
+describe('<WalletInitiatedHeader />', () => {
+  it.only('should match snapshot', () => {
+    const { container } = render();
+
+    expect(container).toMatchSnapshot();
+  });
+});
